+++ assets/javascripts/kitten/components/form/input/password-input/stories.js
@@ -24,6 +24,7 @@ export default {
     iconLabel: 'Show password',
     hiddenIconLabel: 'Hide password',
     name: 'Password',
+    id: 'Password',
   },
 
   argTypes: {