+++ assets/javascripts/kitten/components/navigation/header-nav/components/search-input.js
@@ -134,6 +134,7 @@ export const SearchInput = ({
       onKeyDown={handleKeyDown}
     >
       <TextInputWithButton
+        id="SearchInput"
         size="small"
         rounded
         buttonValue={<SearchIcon />}